ChatBot/requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch
